requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
jieba==0.42.1
scikit-learn==1.3.2
mysqlclient==2.2.0
Pillow==10.1.0 